+++ packages/types/src/ui/stores/grid/index.ts
@@ -1,3 +1,4 @@
 export * from "./columns"
 export * from "./datasource"
 export * from "./table"
+export * from "./view"
